refactor: api types folders
Reorganized into files of shared concerns.
